Add files via upload
I am frequently struck by one of music's particular qualities. Again and again it occurs to me that music puts into plays a unique relationship between the natural and the artificial. We know well how 'natural' music is - it seems culturally universal, it is the quintessentially embodied art form, it seems to be deeply or necessarily enmeshed in the evolution of our species. To sing, to dance, to play with rhythm - these are all so fundamental as to reach into the part of us that is homo sapiens. We also know how 'artificial'